chg: [sysdiagnose] get_parsers and get_analysers move to SysdiagnoseConfig that will ease some analysers, like Timesketch, when querying for parsers.
